pyproject.toml
[tool.poetry]
name = "flasktest"
version = "0.1.0"
description = ""
authors = ["jaarabytes <xavierandole78@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asn1crypto = "1.5.1"
asyncpg = "0.29.0"
blinker = "1.8.2"
click = "8.1.7"
flask = "3.0.3"
flask-sqlalchemy = "3.1.1"
greenlet = "3.0.3"
iniconfig = "2.0.0"
itsdangerous = "2.2.0"
jinja2 = "3.1.4"
markupsafe = "2.1.5"
packaging = "24.1"
pg8000 = "1.31.2"
pluggy = "1.5.0"
pytest = "8.3.2"
pytest-flask = "1.3.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
scramp = "1.4.5"
six = "1.16.0"
sqlalchemy = "2.0.32"
typing-extensions = "4.12.2"
werkzeug = "3.0.4"
celery = "^5.4.0"
eventlet = "^0.36.1"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
